requirements.txt
ebaysdk==2.2.0
python-dotenv==1.0.0
pandas==2.1.4
requests==2.31.0
beautifulsoup4==4.12.2 